Moved `generateApiReport` function into helpers file.
